Force SingleQueryProvider to convert arguments to CamelCase.
Fixes showing invocations for a particular workflow.
