+++ client/src/components/providers/SingleQueryProvider.js
@@ -1,5 +1,6 @@
 import hash from "object-hash";
 import { LastQueue } from "utils/promise-queue";
+import { HasAttributesMixin } from "./utils";
 
 /**
  * Builds a provider that gets its result from a single promise-based query function and
@@ -14,6 +15,7 @@ import { LastQueue } from "utils/promise-queue";
 export const SingleQueryProvider = (lookup, stopRefresh = (result) => false) => {
     const promiseCache = new Map();
     return {
+        mixins: [HasAttributesMixin],
         props: {
             useCache: {
                 type: Boolean,
@@ -74,7 +76,7 @@ export const SingleQueryProvider = (lookup, stopRefresh = (result) => false) => 
                         promiseCache.set(this.cacheKey, lookupPromise);
                     }
                 } else {
-                    lookupPromise = this.queue.enqueue(lookup, this.$attrs);
+                    lookupPromise = this.queue.enqueue(lookup, this.attributes);
                 }
                 lookupPromise.then(
                     (result) => {
